requirements.txt
fastapi
uvicorn[standard]
openai
databases[postgresql]
sqlalchemy
psycopg2-binary
python-dotenv
